Extracted runCommand and added test. WIP
